refactor: ImageLoader 구현 및 DataLoad 관련 작업 DataLoader에 통합
